mod pizzas crud

+++ src/app/home/home.component.ts
@@ -3,6 +3,10 @@ import { ApiService } from '../service/api.service';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import { ViewChild } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { CreatePizzaFormComponent } from '../create-pizza-form/create-pizza-form.component';
+import { EditPizzaFormComponent } from '../edit-pizza-form/edit-pizza-form.component';
+
 
 @Component({
   selector: 'app-home',
@@ -22,7 +26,8 @@ export class HomeComponent implements OnInit, AfterViewInit {
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
 
-  constructor(private apiService: ApiService) { }
+  constructor(private apiService: ApiService, private dialog: MatDialog) { }
+
 
   ngOnInit(): void {
     this.llenarData()
@@ -48,14 +53,14 @@ export class HomeComponent implements OnInit, AfterViewInit {
       this.dataSource.paginator.firstPage(); // Volver a la primera página si el filtro cambia
     }
   }
-  agregarPizza() {
-    this.apiService.crearPizza(this.nuevaPizza).subscribe(
+  agregarPizza(pizza: any) {
+    this.apiService.crearPizza(pizza).subscribe(
       (respuesta) => {
         // Agregar la pizza directamente al dataSource.data
         const newData = this.dataSource.data;
         newData.push(respuesta);
         this.dataSource.data = newData; // Esto es necesario para actualizar la tabla
-
+  
         this.limpiarFormulario();
       },
       (error) => {
@@ -63,10 +68,9 @@ export class HomeComponent implements OnInit, AfterViewInit {
       }
     );
   }
-
-
-  editarPizza() {
-    this.apiService.editarPizza(this.nuevaPizza).subscribe(
+  
+  editarPizza(pizza: any) {
+    this.apiService.editarPizza(pizza).subscribe(
       (respuesta) => {
         // Actualiza la lista de pizzas o recarga los datos
         this.llenarData();
@@ -79,6 +83,7 @@ export class HomeComponent implements OnInit, AfterViewInit {
       }
     );
   }
+  
   startEdit(pizza: any) {
     this.nuevaPizza = { ...pizza };
     this.mostrarForm = true;
@@ -97,17 +102,22 @@ export class HomeComponent implements OnInit, AfterViewInit {
     this.isEditing = false; // Asegurarse de que el estado de edición está desactivado
   }
   eliminarPizza(id: number) {
-    this.apiService.eliminarPizza(id).subscribe({
-      next: () => {
-        // Eliminar la pizza del dataSource.data
-        const newData = this.dataSource.data.filter(pizza => pizza.piz_id !== id);
-        this.dataSource.data = newData; // Actualizar la tabla
-      },
-      error: (error) => {
-        console.error('Error al eliminar pizza', error);
-      }
-    });
+    const confirmarEliminar = window.confirm('¿Seguro que deseas eliminar esta pizza?');
+  
+    if (confirmarEliminar) {
+      this.apiService.eliminarPizza(id).subscribe({
+        next: () => {
+          // Eliminar la pizza del dataSource.data
+          const newData = this.dataSource.data.filter(pizza => pizza.piz_id !== id);
+          this.dataSource.data = newData; // Actualizar la tabla
+        },
+        error: (error) => {
+          console.error('Error al eliminar pizza', error);
+        }
+      });
+    }
   }
+  
 
 
   mostrarFormulario() {
@@ -115,6 +125,32 @@ export class HomeComponent implements OnInit, AfterViewInit {
   }
 
 
-
+  mostrarCrearPizzaDialog() {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+  
+    const dialogRef = this.dialog.open(CreatePizzaFormComponent, dialogConfig);
+    dialogRef.componentInstance.guardarPizza = this.agregarPizza.bind(this); // Pasa la función agregarPizza
+  
+    dialogRef.afterClosed().subscribe(result => {
+      this.llenarData();
+    });
+  }
+  
+  mostrarEditarPizzaDialog(pizza: any) {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.data = pizza;
+  
+    const dialogRef = this.dialog.open(EditPizzaFormComponent, dialogConfig);
+    dialogRef.componentInstance.guardarPizza = this.editarPizza.bind(this); // Pasa la función editarPizza
+  
+    dialogRef.afterClosed().subscribe(result => {
+      this.llenarData();
+    });
+  }
+  
 
 }
